Adding seconds parsing for subsecond intervals

+++ src/scripts/slurm_status.py
@@ -283,15 +283,17 @@ def which(program):
 
 def convert_cpu_to_seconds(cpu_string):
     # The time fields in sacct's output have this format:
-    #   [DD-[hh:]]mm:ss
+    #   [DD-[hh:]]mm:ss.sss
     # Convert that to just seconds.
     elem = re.split('[-:]', cpu_string)
-    secs = int(elem[-1]) + int(elem[-2]) * 60
+
+    # Convert seconds to a float, truncate to int at end
+    secs = float(elem[-1]) + int(elem[-2]) * 60
     if len(elem) > 2:
         secs += int(elem[-3]) * 3600
     if len(elem) > 3:
         secs += int(elem[-4]) * 86400
-    return secs
+    return int(secs)
 
 def get_finished_job_stats(jobid):
     """
